rename `erosion` sub-module to `bank_erosion`
