+++ src/dfastbe/bank_erosion/data_models.py
@@ -196,7 +196,7 @@ class ErosionResults:
     Examples:
         - You can create an instance of the ErosionResults class as follows:
         ```python
-        >>> from dfastbe.erosion.data_models import ErosionResults
+        >>> from dfastbe.bank_erosion.data_models import ErosionResults
         >>> import numpy as np
         >>> erosion_results = ErosionResults(
         ...     eq_erosion_dist=[np.array([0.1, 0.2])],
@@ -220,7 +220,7 @@ class ErosionResults:
         can be set to empty arrays if not needed.
 
         ```python
-        >>> from dfastbe.erosion.data_models import ErosionResults
+        >>> from dfastbe.bank_erosion.data_models import ErosionResults
         >>> import numpy as np
         >>> erosion_results = ErosionResults(
         ...     eq_erosion_dist=[np.array([0.1, 0.2])],
