+++ src/dfastbe/cmd.py
@@ -27,7 +27,7 @@ This file is part of D-FAST Bank Erosion: https://github.com/Deltares/D-FAST_Ban
 """
 from pathlib import Path
 from dfastbe.io import ConfigFile, load_program_texts
-from dfastbe.erosion.bank_erosion import Erosion
+from dfastbe.bank_erosion.bank_erosion import Erosion
 from dfastbe.bank_lines.bank_lines import BankLines
 from dfastbe.gui import main
 from dfastbe import __file__
